+++ CHANGELOG.md
@@ -1,12 +1,17 @@
 # 8.0.0
 * i: perfect-scrollbar update (0.6.12)
 * i: Channel information before stream process to use Youtube channel name in stream name
+* i: Changed websites and streamListFromSetting from object to Map
+* i: Replaced some for...in loops with array.forEach like ones
 * +: Add headers support with the custom Request function
 * +: Add Youtube support
 * +: Re-arranged function to add stream to allow getting IDs from user and video pages with Youtube
 * +: Replaced checkLivesProgress_* function to detect end by promise system (ES6)
+* +: Moved importation real user id
+* +: Now using the promise system to detect end and refresh panel after
 * -: Deleted Request_Get function, no longer needed to avoid copy/paste problem
 * Fix: Unsuccessful channel list requests
+* Fix: Beam importation
 
 # 7.2.6
 * Fix: Offline streams count